requirements.txt
Flask
requests
python-dotenv
urllib3
flask-mail
Flask-mail
APScheduler
schedule